[TD] merge message types files
