+++ crates/sui-core/src/lib.rs
@@ -47,7 +47,6 @@ pub mod transaction_orchestrator;
 mod transaction_outputs;
 pub mod validator_tx_finalizer;
 pub mod verify_indexes;
-mod wait_for_effects_request;
 
 #[cfg(test)]
 #[path = "unit_tests/congestion_control_tests.rs"]
